Added temporary chat history
Added working history function for temporary usage, clears with the
logout.
Moved enum ChatRoomType to his own file.
