+++ src/com/andfchat/core/connection/handler/LeftChannelHandler.java
@@ -28,7 +28,7 @@ import com.andfchat.R;
 import com.andfchat.core.connection.FeedbackListner;
 import com.andfchat.core.connection.ServerToken;
 import com.andfchat.core.data.ChatEntry;
-import com.andfchat.core.data.ChatEntry.ChatEntryType;
+import com.andfchat.core.data.ChatEntryType;
 import com.andfchat.core.data.FlistChar;
 
 /**
