+++ src/com/andfchat/core/data/Chatroom.java
@@ -25,8 +25,6 @@ import java.util.List;
 import roboguice.util.Ln;
 import android.text.Spannable;
 
-import com.andfchat.core.data.ChatEntry.ChatEntryType;
-
 public class Chatroom {
 
     public enum ChatroomType {
@@ -234,5 +232,30 @@ public class Chatroom {
 
     public int getMaxTextLength() {
         return maxTextLength;
+    }
+
+    @Override
+    public int hashCode() {
+        final int prime = 31;
+        int result = 1;
+        result = prime * result + ((channel == null) ? 0 : channel.hashCode());
+        return result;
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj)
+            return true;
+        if (obj == null)
+            return false;
+        if (getClass() != obj.getClass())
+            return false;
+        Chatroom other = (Chatroom) obj;
+        if (channel == null) {
+            if (other.channel != null)
+                return false;
+        } else if (!channel.equals(other.channel))
+            return false;
+        return true;
     }
 }
